requirements.txt
Django
psycopg2-binary
django-environ
Pillow
django-crispy-forms
EOF
pytesseract 
rapidfuzz